🩹fix: id 기본형으로 변경 및 month를 LocalDate로 수정 #9

+++ src/main/java/org/cardGGaduekMainService/card/domain/CardVO.java
@@ -12,10 +12,10 @@ import java.util.Date;
 @NoArgsConstructor
 @Builder
 public class CardVO {
-    private int id;
-    private int customerId;
-    private int accountId;
-    private int cardProductId;
+    private Long id;
+    private Long customerId;
+    private Long accountId;
+    private Long cardProductId;
     private String cardNumber;
     private String cvc;
     private Date cardIssuedDate;
